Implement class tree system
With this commit comes two major breaking changes:
- Abilities can now return 0 for a cost to indicate the ability is free,
and thus will not subtract any points from the player when it is
purchased. Values less than 0 are used internally to indicate specific
reasons as to why an ability cannot be purchased at that time
- Forbidden and required abilities logic has been moved to class trees
instead of individual abilities. The terminology has also changed. The
arrays RequiredByAbilities and ForbidsAbilities are used to denote which
abilities further in the tree depend on or are forbidden from being
purchased if the previous ability is purchased. The arrays
RequiredAbilities and ForbiddenAbilities in RPGAbility are deprecated as
of this commit.