Utilize OverridePickupQuery in RPGAbility for weapon pickups in RPGRules::OverridePickupQuery and allow abilities to modify weapons after they have been granted
This introduces a breaking backwards-compatibility change with TitanRPG.
RPGAbility::ModifyGrantedWeapon has been renamed to
OverrideGrantedWeapon, but retains its original functionality.
RPGAbility::ModifyGrantedWeaponAmmo has also been renamed to
OverrideGrantedWeaponAmmo. The newly added
RPGAbility::ModifyGrantedWeapon takes two parameters: the weapon and the
modifier of the weapon that was just received. This allows an ability to
modify a weapon after it has actually been granted.