Add GetHUDOptionCosts function to RPGArtifact
This can be used to return an array of icon materials, costs, and
whether or not this particular cost can be afforded in selection
artifacts. This allows the selection artifact dialog to display a series
of costs and icons if the selection artifact requires more than
or something other than adrenaline. If this function returns an empty
array (default), the standard adrenaline icon and the value of
GetOptionCost(i) will be used for all option costs.