Implement new build system
The new build process is handled by build.sh which can be invoked
manually or by using `make` or `make compile`. The build script also
implements the usage of `gpp` as a preprocessor running in CPP mode. The
build script will copy the entire tree of Classes/ to .Classes/ and
run `gpp` against all *.uc files in Classes/. Upon the build script
either finishing successfully or on failure, it will remove the entire
Classes/ tree and restore the old one, unless the `-n|--no-restore` option is
specified. Additionally, the `-s|--skip-preprocess` option can be used
to skip the preprocessing phase, for whatever reason. Lastly, when the
tree has been preprocessed it will be marked by the creation of a
.preprocessed file, which, when detected by the script, will cause the
preprocessing phase to be skipped. A warning will be printed to the user
if this occurs. `make clean` has been extended to not only force
recompilation even if the source tree is unchanged, but also to restore
the original Classes/ tree if the `-n|--no-restore` option was
previously used.