Implement total hack to fix ONSWeaponPawn and ONSWeapon not rotating properly on clients
This appears to happen because of a typo in ONSWeapon.cpp:295. When the
player enters an ONSWeaponPawn, its corresponding ONSWeapon's Owner gets
set to the ONSWeaponPawn. When the player leaves, the Owner of the
ONSWeapon is maintained, even when other players enter it. The 'if'
statement on line 295 in ONSWeapon.cpp checks if 'Owner' is not NULL,
which will fail the entire statement once this happens. It would appear
as though this check to 'Owner' should be '!bNetOwner' instead. By
forcibly setting and resetting the ONSWeapon's Owner, we can workaround
this