Add rudimentary server target in Makefile
This target spins up a temporary test server on port 7777 (default). The
server is alive for as long as it does not crash or is interrupted with
CTRL+C. The properties of the server can be edited in the included
server.ini and in the startup script server.sh. Additionally, a
rudimentary redirect server will automatically be started along with the
game server that will serve up files out of the server's usual
directory tree. This system allows for quickly spinning up a test server
with all mutators and settings present and without any further
configuration necessary.