mut: make random weapon modifier generator fair
If a random weapon modifier was requested but the one chosen is not
allowed on the weapon, the next modifier which was allowed would be
chosen. This results in weapon modifiers at the beginning of the list to
have an advantage. Instead, just retry the entire routine until we
either get a modifier or give up.